--- styleguide/src/Layout/Box/BoxHeader.tsx
@@ -1,62 +1,0 @@
-import React, { useContext } from 'react'
-
-import { TabsProps } from '../../Components/Tabs'
-import { SharedContext, defaultPaddingVariants } from './utils'
-
-export const BoxHeader = React.memo(
-  ({ children, title, subtitle, Tabs }: BoxHeaderProps) => {
-    const sharedProps = useContext(SharedContext)
-    const { variant } = sharedProps
-
-    const hasTabs = !!Tabs
-
-    return (
-      <div
-        className={`box-header border-b border-card-stroke ${
-          defaultPaddingVariants[variant]
-        } ${hasTabs ? '!pb-0' : ''}`}
-      >
-        <div
-          className={`flex justify-between items-center ${
-            !subtitle ? 'py-1' : ''
-          }`}
-        >
-          <div className="flex-1 min-w-0 mr-2">
-            {title && (
-              <h3 className={`tracking-3 text-xl font-semibold break-words`}>
-                {title}
-              </h3>
-            )}
-            {subtitle && (
-              <h4 className="tracking-4 text-base text-on-base-2 leading-6 break-words mt-1">
-                {subtitle}
-              </h4>
-            )}
-          </div>
-          {children}
-        </div>
-        {hasTabs && <div className={`mt-3 -mb-px`}>{Tabs}</div>}
-      </div>
-    )
-  }
-)
-
-export interface BoxHeaderProps {
-  /**
-   * Title of the Box
-   */
-  title: string | React.ReactNode
-  /**
-   * Subtitle of the Box
-   */
-  subtitle?: string | React.ReactNode
-  /**
-   * React children, use to render actions in Header
-   * Also support render prop
-   */
-  children?: React.ReactNode | ((props: BoxHeaderProps) => React.ReactNode)
-  /**
-   * Use only <Tabs /> component
-   */
-  Tabs?: React.ReactElement<TabsProps>
-}
