+++ styleguide/src/Layout/Box/Components/BoxContent.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Story, Meta } from '@storybook/react'
 
-import { Box } from './Box'
+import { Box } from '../Box'
 import { BoxContentProps } from './BoxContent'
 
 export default {
