+++ styleguide/src/Layout/Box/Components/BoxHeader.stories.tsx
@@ -1,12 +1,20 @@
 import React from 'react'
 import { Story, Meta } from '@storybook/react'
 
-import { Box, BoxProps } from './Box'
+import { Box, BoxProps } from '../Box'
 import { BoxHeaderProps } from './BoxHeader'
 
 export default {
   title: 'Layout/Box/Header',
   component: Box.Header,
+  argTypes: {
+    Tabs: {
+      control: false
+    },
+    children: {
+      control: false
+    }
+  },
   args: {
     title: 'Box Title',
     subtitle: 'Box SubTitle',
